+++ packages/agent/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Added
+
+- Added `Agent.setQueuedMessageGrouping()` so hosts can deliver adjacent companion records and their prompt together in `one-at-a-time` queue modes ([#11618](https://github.com/can1357/oh-my-pi/pull/11618) by [@andrebrait](https://github.com/andrebrait)).
+
 ## [18.1.17] - 2026-09-10
 
 ### Changed